Re-structuring goapp
