+++ app/models/models.go
@@ -1,0 +1,3 @@
+package models
+
+const DB_POST_TABLE = "PostNew"
